Moved DevicePanel MVC to device.py
